#86 #87 #88

+++ src/main/java/com/springboot/manufacture_history/entity/ManuFactureHistory.java
@@ -36,7 +36,7 @@ public class ManuFactureHistory {
     private BigDecimal unitPrice;
 
     @Column(nullable = false)
-    private Long qty;
+    private int qty;
 
     @Column(nullable = false)
     private LocalDateTime receiveDate;
